% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSbinomialestimationWrapper.R
\name{LSbinomialestimation}
\alias{LSbinomialestimation}
\title{Binomial Estimation}
\usage{
LSbinomialestimation(
  data = NULL,
  version = "1",
  colorPalette = "colorblind",
  dataCountsFailures = 0,
  dataCountsSuccesses = 0,
  dataInputType = "counts",
  dataSequenceFailures = list(types = list(), value = list()),
  dataSequenceSequenceOfObservations = "",
  dataSequenceSuccesses = list(types = list(), value = list()),
  dataSummary = TRUE,
  dataVariableFailures = list(types = list(), value = list()),
  dataVariableSelected = list(types = list(), value = ""),
  dataVariableSuccesses = list(types = list(), value = list()),
  introductoryText = FALSE,
  models = list(),
  plotHeight = 320,
  plotWidth = 480,
  posteriorDistributionPloPriorDistribution = FALSE,
  posteriorDistributionPlot = FALSE,
  posteriorDistributionPlotIndividualCi = FALSE,
  posteriorDistributionPlotIndividualCiBf = 1,
  posteriorDistributionPlotIndividualCiLower = 0.25,
  posteriorDistributionPlotIndividualCiMass = 0.95,
  posteriorDistributionPlotIndividualCiType = "central",
  posteriorDistributionPlotIndividualCiUpper = 0.75,
  posteriorDistributionPlotIndividualPointEstimate = FALSE,
  posteriorDistributionPlotIndividualPointEstimateType = "mean",
  posteriorDistributionPlotObservedProportion = FALSE,
  posteriorDistributionPlotType = "overlying",
  posteriorPredictionDistributionPlot = FALSE,
  posteriorPredictionDistributionPlotAsSampleProportion = FALSE,
  posteriorPredictionDistributionPlotIndividualCi = FALSE,
  posteriorPredictionDistributionPlotIndividualCiLower = 0,
  posteriorPredictionDistributionPlotIndividualCiMass = 0.95,
  posteriorPredictionDistributionPlotIndividualCiType = "central",
  posteriorPredictionDistributionPlotIndividualCiUpper = 1,
  posteriorPredictionDistributionPlotIndividualPointEstimate = FALSE,
  posteriorPredictionDistributionPlotIndividualPointEstimateType = "mean",
  posteriorPredictionDistributionPlotPredictionsTable = FALSE,
  posteriorPredictionDistributionPlotType = "overlying",
  posteriorPredictionNumberOfFutureTrials = 10,
  posteriorPredictionSummaryTable = FALSE,
  posteriorPredictionSummaryTablePointEstimate = "mean",
  priorAndPosteriorPointEstimate = "mean",
  priorDistributionPlot = FALSE,
  priorDistributionPlotIndividualCi = FALSE,
  priorDistributionPlotIndividualCiLower = 0.25,
  priorDistributionPlotIndividualCiMass = 0.95,
  priorDistributionPlotIndividualCiType = "central",
  priorDistributionPlotIndividualCiUpper = 0.75,
  priorDistributionPlotIndividualPointEstimate = FALSE,
  priorDistributionPlotIndividualPointEstimateType = "mean",
  priorDistributionPlotType = "overlying",
  sequentialAnalysisIntervalEstimatePlot = FALSE,
  sequentialAnalysisIntervalEstimatePlotLower = 0.25,
  sequentialAnalysisIntervalEstimatePlotType = "overlying",
  sequentialAnalysisIntervalEstimatePlotUpdatingTable = FALSE,
  sequentialAnalysisIntervalEstimatePlotUpper = 0.75,
  sequentialAnalysisPointEstimatePlot = FALSE,
  sequentialAnalysisPointEstimatePlotCi = FALSE,
  sequentialAnalysisPointEstimatePlotCiBf = 1,
  sequentialAnalysisPointEstimatePlotCiMass = 0.95,
  sequentialAnalysisPointEstimatePlotCiType = "central",
  sequentialAnalysisPointEstimatePlotType = "mean",
  sequentialAnalysisPointEstimatePlotUpdatingTable = FALSE,
  sequentialAnalysisPosteriorUpdatingTable = FALSE,
  sequentialAnalysisStackedDistributionsPlot = FALSE
)
}
\description{
Binomial Estimation
}
