% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSBuffonsneedlemanipulationWrapper.R
\name{LSBuffonsneedlemanipulation}
\alias{LSBuffonsneedlemanipulation}
\title{Manipulating Buffon's Needle}
\usage{
LSBuffonsneedlemanipulation(
  data = NULL,
  version = "1",
  ciLevel = 0.95,
  highlight = FALSE,
  lengthToDistanceProportion = 80,
  max = 3.2,
  min = 3,
  numberOfCrosses = 50,
  numberOfThrows = 100,
  plotHeight = 320,
  plotWidth = 480,
  priorAlpha = 1,
  priorBeta = 1,
  priorPosteriorPi = TRUE,
  priorPosteriorPiCi = FALSE,
  priorPosteriorPiLegend = FALSE,
  priorPosteriorProportion = FALSE,
  priorPosteriorProportionCi = FALSE,
  priorPosteriorProportionLegend = FALSE
)
}
\description{
Manipulating Buffon's Needle
}
