% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSgameofchanceWrapper.R
\name{LSgameofchance}
\alias{LSgameofchance}
\title{Game of Chance}
\usage{
LSgameofchance(
  data = NULL,
  version = "0.97.1",
  ci = TRUE,
  numberOfSimulatedGames = 500,
  players = list(list(levels = list("p(win 1 point)", "Points gained"), name = " 1",
    values = list(1, 0)), list(levels = list("p(win 1 point)", "Points gained"), name =
    " 2", values = list(1, 0))),
  plotHeight = 320,
  plotWidth = 480,
  pointsToWin = 2
)
}
\description{
Game of Chance
}
