% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSbinaryclassificationWrapper.R
\name{LSbinaryclassification}
\alias{LSbinaryclassification}
\title{Binary Classification}
\usage{
LSbinaryclassification(
  data = NULL,
  version = "0.97.1",
  alluvialPlot = FALSE,
  areaPlot = FALSE,
  burnin = 500,
  chains = 4,
  ci = TRUE,
  ciLevel = 0.95,
  colorFalseNegative = "red",
  colorFalsePositive = "darkorange",
  colorTrueNegative = "steelblue",
  colorTruePositive = "darkgreen",
  computeResults = TRUE,
  confusionMatrix = FALSE,
  confusionMatrixAdditionalInfo = TRUE,
  confusionMatrixType = "text",
  estimatesPlot = FALSE,
  estimatesPlotAccuracy = FALSE,
  estimatesPlotFalseDiscoveryRate = FALSE,
  estimatesPlotFalseNegative = FALSE,
  estimatesPlotFalseNegativeRate = FALSE,
  estimatesPlotFalseOmissionRate = FALSE,
  estimatesPlotFalsePositive = FALSE,
  estimatesPlotFalsePositiveRate = FALSE,
  estimatesPlotNegativePredictiveValue = FALSE,
  estimatesPlotPositivePredictiveValue = FALSE,
  estimatesPlotPrevalence = TRUE,
  estimatesPlotSensitivity = TRUE,
  estimatesPlotSpecificity = TRUE,
  estimatesPlotTrueNegative = FALSE,
  estimatesPlotTruePositive = FALSE,
  falseNegative = 0,
  falsePositive = 0,
  iconPlot = FALSE,
  inputType = "pointEstimates",
  introductoryText = FALSE,
  labels = list(types = list(), value = ""),
  marker = list(types = list(), value = ""),
  negativeTests = 0,
  orderConstraint = TRUE,
  plotEstimatesType = "halfEye",
  plotHeight = 320,
  plotWidth = 480,
  positiveTests = 0,
  ppvNpvPlot = FALSE,
  prPlot = FALSE,
  prPlotPosteriorRealizations = FALSE,
  prPlotPosteriorRealizationsNumber = 100,
  predictiveValuesByPrevalence = FALSE,
  prevalence = 0.1,
  priorPosterior = FALSE,
  priorPrevalenceAlpha = 1,
  priorPrevalenceBeta = 9,
  priorSensitivityAlpha = 8,
  priorSensitivityBeta = 2,
  priorSpecificityAlpha = 8,
  priorSpecificityBeta = 2,
  probabilityPositivePlot = FALSE,
  probabilityPositivePlotEntireDistribution = FALSE,
  rocPlot = FALSE,
  rocPlotPosteriorRealizations = FALSE,
  rocPlotPosteriorRealizationsNumber = 100,
  samples = 10000,
  seed = 1,
  sensitivity = 0.8,
  setSeed = FALSE,
  signalDetectionPlot = FALSE,
  specificity = 0.8,
  statistics = TRUE,
  statisticsAdditional = FALSE,
  testCharacteristicsPlot = FALSE,
  thinning = 1,
  threshold = 0,
  tocPlot = FALSE,
  trueNegative = 0,
  truePositive = 0,
  updatePrevalence = TRUE,
  varyingThresholdBurnin = 500,
  varyingThresholdChains = 2,
  varyingThresholdSamples = 1000,
  varyingThresholdThinning = 1
)
}
\description{
Binary Classification
}
