% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSbinomialtestingWrapper.R
\name{LSbinomialtesting}
\alias{LSbinomialtesting}
\title{Binomial Testing}
\usage{
LSbinomialtesting(
  data = NULL,
  version = "0.97.1",
  colorPalette = "colorblind",
  dataCountsFailures = 0,
  dataCountsSuccesses = 0,
  dataInputType = "counts",
  dataSequenceFailures = list(types = list(), value = list()),
  dataSequenceSequenceOfObservations = "",
  dataSequenceSuccesses = list(types = list(), value = list()),
  dataSummary = TRUE,
  dataVariableFailures = list(types = list(), value = list()),
  dataVariableSelected = list(types = list(), value = ""),
  dataVariableSuccesses = list(types = list(), value = list()),
  introductoryText = FALSE,
  models = list(),
  plotHeight = 320,
  plotWidth = 480,
  posteriorDistributionPlot = FALSE,
  posteriorDistributionPlotConditionalCi = FALSE,
  posteriorDistributionPlotConditionalCiBf = 1,
  posteriorDistributionPlotConditionalCiLower = 0.25,
  posteriorDistributionPlotConditionalCiMass = 0.95,
  posteriorDistributionPlotConditionalCiType = "central",
  posteriorDistributionPlotConditionalCiUpper = 0.75,
  posteriorDistributionPlotConditionalPointEstimate = FALSE,
  posteriorDistributionPlotConditionalPointEstimateType = "mean",
  posteriorDistributionPlotJointType = "overlying",
  posteriorDistributionPlotMarginalCi = FALSE,
  posteriorDistributionPlotMarginalCiBf = 1,
  posteriorDistributionPlotMarginalCiLower = 0.25,
  posteriorDistributionPlotMarginalCiMass = 0.95,
  posteriorDistributionPlotMarginalCiType = "central",
  posteriorDistributionPlotMarginalCiUpper = 0.75,
  posteriorDistributionPlotMarginalPointEstimate = FALSE,
  posteriorDistributionPlotMarginalPointEstimateType = "mean",
  posteriorDistributionPlotObservedProportion = FALSE,
  posteriorDistributionPlotType = "conditional",
  posteriorPredictionDistributionPlot = FALSE,
  posteriorPredictionDistributionPlotAsSampleProportion = FALSE,
  posteriorPredictionDistributionPlotConditionalCi = FALSE,
  posteriorPredictionDistributionPlotConditionalCiLower = 0,
  posteriorPredictionDistributionPlotConditionalCiMass = 0.95,
  posteriorPredictionDistributionPlotConditionalCiType = "central",
  posteriorPredictionDistributionPlotConditionalCiUpper = 1,
  posteriorPredictionDistributionPlotConditionalPointEstimate = FALSE,
  posteriorPredictionDistributionPlotConditionalPointEstimateType = "mean",
  posteriorPredictionDistributionPlotJoinType = "overlying",
  posteriorPredictionDistributionPlotMarginalCi = FALSE,
  posteriorPredictionDistributionPlotMarginalCiLower = 0,
  posteriorPredictionDistributionPlotMarginalCiMass = 0.95,
  posteriorPredictionDistributionPlotMarginalCiType = "central",
  posteriorPredictionDistributionPlotMarginalCiUpper = 1,
  posteriorPredictionDistributionPlotMarginalPointEstimate = FALSE,
  posteriorPredictionDistributionPlotMarginalPointEstimateType = "mean",
  posteriorPredictionDistributionPlotPredictionsTable = FALSE,
  posteriorPredictionDistributionPlotType = "conditional",
  posteriorPredictionNumberOfFutureTrials = 10,
  posteriorPredictionSummaryTable = FALSE,
  posteriorPredictionSummaryTablePointEstimate = "mean",
  priorAndPosteriorDistributionPlot = FALSE,
  priorAndPosteriorDistributionPlotObservedProportion = FALSE,
  priorAndPosteriorDistributionPlotType = "conditional",
  priorDistributionPlot = FALSE,
  priorDistributionPlotConditionalCi = FALSE,
  priorDistributionPlotConditionalCiLower = 0.25,
  priorDistributionPlotConditionalCiMass = 0.95,
  priorDistributionPlotConditionalCiType = "central",
  priorDistributionPlotConditionalCiUpper = 0.75,
  priorDistributionPlotConditionalPointEstimate = FALSE,
  priorDistributionPlotConditionalPointEstimateType = "mean",
  priorDistributionPlotJointType = "overlying",
  priorDistributionPlotMarginalCi = FALSE,
  priorDistributionPlotMarginalCiLower = 0.25,
  priorDistributionPlotMarginalCiMass = 0.95,
  priorDistributionPlotMarginalCiType = "central",
  priorDistributionPlotMarginalCiUpper = 0.75,
  priorDistributionPlotMarginalPointEstimate = FALSE,
  priorDistributionPlotMarginalPointEstimateType = "mean",
  priorDistributionPlotType = "conditional",
  priorPredictivePerformanceAccuracyPlot = FALSE,
  priorPredictivePerformanceAccuracyPlotType = "conditional",
  priorPredictivePerformanceBfComparison = "vs",
  priorPredictivePerformanceBfType = "BF10",
  priorPredictivePerformanceBfVsHypothesis = "",
  priorPredictivePerformanceDistributionPlot = FALSE,
  priorPredictivePerformanceDistributionPlotConditionalCi = FALSE,
  priorPredictivePerformanceDistributionPlotConditionalCiLower = 0,
  priorPredictivePerformanceDistributionPlotConditionalCiMass = 0.95,
  priorPredictivePerformanceDistributionPlotConditionalCiType = "central",
  priorPredictivePerformanceDistributionPlotConditionalCiUpper = 1,
  priorPredictivePerformanceDistributionPlotConditionalPointEstimate = FALSE,
  priorPredictivePerformanceDistributionPlotConditionalPointEstimateType = "mean",
  priorPredictivePerformanceDistributionPlotJoinType = "overlying",
  priorPredictivePerformanceDistributionPlotMarginalCi = FALSE,
  priorPredictivePerformanceDistributionPlotMarginalCiLower = 0,
  priorPredictivePerformanceDistributionPlotMarginalCiMass = 0.95,
  priorPredictivePerformanceDistributionPlotMarginalCiType = "central",
  priorPredictivePerformanceDistributionPlotMarginalCiUpper = 1,
  priorPredictivePerformanceDistributionPlotMarginalPointEstimate = FALSE,
  priorPredictivePerformanceDistributionPlotMarginalPointEstimateType = "mean",
  priorPredictivePerformanceDistributionPlotObservedNumberOfSuccessess = FALSE,
  priorPredictivePerformanceDistributionPlotPredictionsTable = FALSE,
  priorPredictivePerformanceDistributionPlotType = "conditional",
  sequentialAnalysisPredictivePerformancePlot = FALSE,
  sequentialAnalysisPredictivePerformancePlotBfComparison = "inclusion",
  sequentialAnalysisPredictivePerformancePlotBfType = "BF10",
  sequentialAnalysisPredictivePerformancePlotBfVsHypothesis = "",
  sequentialAnalysisPredictivePerformancePlotType = "conditional",
  sequentialAnalysisPredictivePerformancePlotUpdatingTable = FALSE
)
}
\description{
Binomial Testing
}
