% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LSgameofskillWrapper.R
\name{LSgameofskill}
\alias{LSgameofskill}
\title{Game of Skill}
\usage{
LSgameofskill(
  data = NULL,
  version = "0.97.1",
  ci = TRUE,
  numberOfSimulatedGames = 500,
  players = list(list(levels = list("Prior skill parameter", "Points gained"), name =
    " 1", values = list(1, 0)), list(levels = list("Prior skill parameter",
    "Points gained"), name = " 2", values = list(1, 0))),
  plotHeight = 320,
  plotWidth = 480,
  pointsToWin = 2
)
}
\description{
Game of Skill
}
